Split service into multiple modules
