+++ account-service/build.gradle
@@ -1,5 +1,6 @@
 plugins {
-    id "account.service-conventions"
+    id("account.service-conventions")
+    id("java-test-fixtures")
 }
 
 dependencies {
@@ -11,19 +12,16 @@ dependencies {
 
     implementation("org.msgpack:msgpack-core:0.8.21")
 
-    implementation("org.flywaydb:flyway-core:6.5.6")
-
     implementation("ch.qos.logback:logback-classic:1.2.3")
     implementation("io.micrometer:micrometer-registry-prometheus:1.5.5")
     implementation("io.zipkin.brave:brave-instrumentation-sparkjava:5.12.5")
     implementation("io.zipkin.brave:brave-instrumentation-p6spy:5.12.5")
     implementation("io.zipkin.reporter2:zipkin-sender-urlconnection:2.15.1")
+    implementation(project(":eventstore"))
 
     implementation("com.zaxxer:HikariCP:3.4.5")
-    implementation("mysql:mysql-connector-java:8.0.21") {
-        exclude(group: "com.google.protobuf", module: "protobuf-java")
-    }
-    implementation("org.postgresql:postgresql:42.2.16")
+    implementation(project(":eventstore-mysql"))
+    implementation(project(":eventstore-postgres"))
 
     testFixturesApi("org.junit.jupiter:junit-jupiter:5.6.2")
     testFixturesApi("org.assertj:assertj-core:3.17.2")
@@ -31,13 +29,6 @@ dependencies {
     testImplementation("com.tngtech.archunit:archunit-junit5:0.14.1")
     testImplementation("io.rest-assured:rest-assured:4.3.1")
     testImplementation("com.h2database:h2:1.4.200")
-
-    integrationTestImplementation("org.testcontainers:mysql:1.14.3") {
-        exclude(group: "org.slf4j", module: "slf4j-api")
-    }
-    integrationTestImplementation("org.testcontainers:postgresql:1.14.3") {
-        exclude(group: "org.slf4j", module: "slf4j-api")
-    }
 
     e2eTestImplementation("org.junit.jupiter:junit-jupiter:5.6.2")
     e2eTestImplementation("org.assertj:assertj-core:3.17.2")
@@ -59,6 +50,8 @@ application {
 }
 
 tasks.named("e2eTest") {
+    it.dependsOn(":eventstore-postgres:integrationTest")
+    it.dependsOn(":eventstore-mysql:integrationTest")
     it.inputs.files(jar.archiveFile, file('Dockerfile'), fileTree("../infrastructure"))
             .withPathSensitivity(PathSensitivity.RELATIVE)
 }
