+++ eventstore-mysql/src/main/java/lt/rieske/accounts/eventstore/MySqlEventStoreFactory.java
@@ -1,0 +1,29 @@
+package lt.rieske.accounts.eventstore;
+
+import com.mysql.cj.jdbc.MysqlDataSource;
+
+import javax.sql.DataSource;
+import java.util.function.Function;
+
+public class MySqlEventStoreFactory {
+
+    public static BlobEventStore mysqlEventStore(String jdbcUrl, String username, String password, Function<DataSource, DataSource> initializer) {
+        return mysqlEventStore(mysqlDataSource(jdbcUrl, username, password), initializer);
+    }
+
+    static BlobEventStore mysqlEventStore(DataSource dataSource, Function<DataSource, DataSource> initializer) {
+        DataSourceConfiguration.migrateDatabase(dataSource, "db/mysql");
+        return new MySqlEventStore(initializer.apply(dataSource));
+    }
+
+    private static DataSource mysqlDataSource(String jdbcUrl, String username, String password) {
+        var dataSource = new MysqlDataSource();
+        dataSource.setUrl(jdbcUrl);
+        dataSource.setUser(username);
+        dataSource.setPassword(password);
+
+        DataSourceConfiguration.waitForDatabaseToBeAvailable(dataSource);
+
+        return dataSource;
+    }
+}
