+++ eventstore-mysql/src/main/resources/db/mysql/V001__init.sql
