+++ eventstore/build.gradle
@@ -1,0 +1,10 @@
+plugins {
+    id("account.java-conventions")
+}
+
+dependencies {
+    implementation("org.slf4j:slf4j-api:1.7.25")
+    implementation("org.flywaydb:flyway-core:6.5.6")
+
+    testFixturesApi(testFixtures(project(":account-service")))
+}
