+++ eventstore/src/main/java/lt/rieske/accounts/eventstore/SerializedEvent.java
