+++ gradle.properties
@@ -1,1 +1,2 @@
+org.gradle.parallel=true
 org.gradle.vfs.watch=true
