+++ gradle/build-logic/src/main/groovy/account.java-conventions.gradle
@@ -1,6 +1,8 @@
 plugins {
     id "java"
+    id "java-test-fixtures"
     id "jacoco"
+    id "account.integration-tests"
 }
 
 java {
@@ -19,7 +21,7 @@ tasks.withType(JavaCompile).configureEach {
 
 tasks.withType(Test).configureEach {
     useJUnitPlatform()
-    maxParallelForks = 8
+    maxParallelForks = 4
     jvmArgs += "--enable-preview"
 }
